float convert
git-svn-id: https://scm9.informatik.uni-erlangen.de/simanies/svn/Optimization@80 bfabd5f7-08ac-4c2d-9fe7-a7df3bb4fc23

+++ API/testMLib/main.cpp
@@ -11,9 +11,9 @@ int main()
     /*OptState * S =  Opt_NewState();
     uint64_t dims[] = {1024,512};
     
-    ImageBinding * A = Opt_ImageBind(S,NULL,1024,sizeof(double));
-    ImageBinding * x = Opt_ImageBind(S,NULL,1,sizeof(double));
-    ImageBinding * b = Opt_ImageBind(S,NULL,1,sizeof(double));
+    ImageBinding * A = Opt_ImageBind(S,NULL,1024,sizeof(float));
+    ImageBinding * x = Opt_ImageBind(S,NULL,1,sizeof(float));
+    ImageBinding * b = Opt_ImageBind(S,NULL,1,sizeof(float));
     
     cout << "start Opt_ProblemDefine" << endl;
     Problem * prob = Opt_ProblemDefine(S,"example.t","gradientdescent",NULL);
